% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summaries.R
\name{calculate_character_summaries}
\alias{calculate_character_summaries}
\title{Calculate columnwise summaries for a data frame.}
\usage{
calculate_character_summaries(dataframe,
  summaries = default_character_summaries, parallel = FALSE,
  mc.cores = parallel::detectCores())
}
\arguments{
\item{dataframe}{data.frame. The data you want to summarize.}

\item{summaries}{list. A list of summary functions. Each function must take
as an input one numeric vector, and output a numeric vector of length 1.}

\item{parallel}{logical. Should we use \code{mclapply} instead of \code{lapply}?}

\item{mc.cores}{numeric. To be passed into \code{mclapply}.}
}
\description{
Calculate columnwise summaries for a data frame.
}

