% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summaries.R
\name{calculate_numeric_summaries}
\alias{calculate_numeric_summaries}
\title{Calculate columnwise summaries for a data frame.}
\usage{
calculate_numeric_summaries(dataframe, summaries = default_numeric_summaries,
  parallel = FALSE, mc.cores = parallel::detectCores())
}
\arguments{
\item{dataframe}{data.frame. The data you want to summarize.}

\item{summaries}{list. A list of summary functions. Each function must take
as an input one numeric vector, and output a numeric vector of length 1.}

\item{parallel}{logical. Should we use \code{mclapply} instead of \code{lapply}?}

\item{mc.cores}{numeric. To be passed into \code{mclapply}.}
}
\value{
data.frame, with a column called \code{variables} with the column names
  from the data frame, and a column for each summary function.
}
\description{
Calculate columnwise summaries for a data frame.
}

