% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_distributions.R
\name{compare_character_distributions}
\alias{compare_character_distributions}
\title{Compare the distribution of common fields across two data frames.}
\usage{
compare_character_distributions(challenger, baseline,
  summaries = default_character_summaries,
  tests = default_character_tests(discrete_tolerance =
  getOption("vardist.character_discrete_tolerance", 2), percent_tolerance =
  getOption("vardist.character_percent_tolerance", 0.1), uniques_threshold =
  getOption("vardist.character_uniques_threshold", 100), uniques_tolerance =
  getOption("vardist.character_uniques_tolerance", 10)), parallel = FALSE,
  mc.cores = parallel::detectCores())
}
\arguments{
\item{challenger}{data.frame.}

\item{baseline}{data.frame.}

\item{summaries}{list. A named list of summary functions. Each function must take
as an input one numeric vector, and output a numeric vector of length 1.}

\item{tests}{list. A named list of functions that return \code{TRUE} or \code{FALSE}, and take in
columns or summary statistics fom the \code{challenger} and the \code{baseline}.}

\item{parallel}{logical. Should we use \code{mclapply} instead of \code{lapply}?}

\item{mc.cores}{numeric. To be passed into \code{mclapply}.}
}
\description{
Compare the distribution of common fields across two data frames.
}
\seealso{
compare_numeric_distributions
}

