% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_distributions.R
\name{compare_numeric_distributions}
\alias{compare_numeric_distributions}
\title{Compare the distribution of common fields across two data frames.}
\usage{
compare_numeric_distributions(challenger, baseline,
  summaries = default_numeric_summaries,
  tests = default_numeric_tests(tolerance =
  getOption("vardist.numeric_summary_tolerance", 0.1), ks_test_threshold =
  getOption("vardist.ks_test_threshold", 0.5)), parallel = FALSE,
  mc.cores = parallel::detectCores())
}
\arguments{
\item{challenger}{data.frame.}

\item{baseline}{data.frame.}

\item{summaries}{list. A named list of summary functions. Each function must take
as an input one numeric vector, and output a numeric vector of length 1.}

\item{tests}{list. A named list of functions that return \code{TRUE} or \code{FALSE}, and take in
columns or summary statistics fom the \code{challenger} and the \code{baseline}.}

\item{parallel}{logical. Should we use \code{mclapply} instead of \code{lapply}?}

\item{mc.cores}{numeric. To be passed into \code{mclapply}.}
}
\value{
a list with three data frames - the columnwise summaries for the \code{challenger},
  the columnwise summaries for the \code{baseline}, and a report with the results of the
  \code{tests}.
}
\description{
Typically you have a data set whose integrity is unknown, and you want to compare
it to a data set whose reliability has already been established by other means.
With this function, you can compare the uncertain data set (the "challenger") to
the certain one (the "baseline") and see if they have similar enough distributions.
}
\seealso{
calculate_summaries, generate_report
}

