% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Compare summary statistics for two datasets, given the summaries and the datasets.}
\usage{
generate_report(challenger_stats, baseline_stats, challenger_df, baseline_df,
  tests, parallel = FALSE, mc.cores = parallel::detectCores())
}
\arguments{
\item{challenger_stats}{data.frame. Summaries for the challenger data frame,
created by \code{calculate_summaries}.}

\item{baseline_stats}{data.frame. Summaries for the baseline data frame,
created by \code{calculate_summaries}.}

\item{challenger_df}{data.frame.}

\item{baseline_df}{data.frame.}

\item{tests}{list. A named list of functions that compare the challenger and
the baseline. See above for details.}

\item{parallel}{logical. Should we use \code{mclapply} instead of \code{lapply}?}

\item{mc.cores}{numeric. To be passed into \code{mclapply}.}
}
\value{
a data.frame of logical values showing if each test passed or failed
  for each column.
}
\description{
In the list \code{tests}, some elements may be named after some of the
summaries (which at this point are column names in \code{baseline_stats}).
}
\details{
If an element in \code{tests} is named after a summary, then the element is
expected to be a function that takes in the baseline statistic, the challenger
statistic, and the respective columns. Otherwise, it is expected to take in
the two columns.
}

