% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report_from_summaries}
\alias{generate_report_from_summaries}
\title{Compare summary statistics for two datasets, given only the summaries.}
\usage{
generate_report_from_summaries(challenger_stats, baseline_stats, tests)
}
\arguments{
\item{challenger_stats}{data.frame. Summaries for the challenger data frame,
created by \code{calculate_summaries}.}

\item{baseline_stats}{data.frame. Summaries for the baseline data frame,
created by \code{calculate_summaries}.}

\item{tests}{list. A named list of functions that compare the challenger and
the baseline. See above for details.}

\item{challenger_df}{data.frame.}

\item{baseline_df}{data.frame.}
}
\value{
a data.frame of logical values showing if each test passed or failed
  for each column.
}
\description{
Compare summary statistics for two datasets, given only the summaries.
}

